[refact]: reorganized the folder structure
