+++ src/lib/php/Dao/FolderDao.php
@@ -264,7 +264,7 @@ WHERE fc.parent_fk = $1 AND fc.foldercontents_mode = " . self::MODE_UPLOAD . " A
 
   public function createFolder($folderName, $folderDescription, $parentId)
   {
-    $folderId = $this->dbManager->insertTableRow("folder", array("folder_pk" => 5,"folder_name" => $folderName, "user_fk" => Auth::getUserId(), "folder_desc" => $folderDescription), null,'folder_pk');
+    $folderId = $this->dbManager->insertTableRow("folder", array("folder_name" => $folderName, "user_fk" => Auth::getUserId(), "folder_desc" => $folderDescription), null, 'folder_pk');
     $this->insertFolderContents($parentId, self::MODE_FOLDER, $folderId);
     return $folderId;
   }
