+++ src/lib/php/Dao/test/FolderDaoTest.php
@@ -24,14 +24,16 @@ class FolderDaoTest extends \PHPUnit\Framework\TestCase
 
   protected function setUp() : void
   {
+    global $SysConf;
+    $SysConf = [''];
     $this->testDb = new TestPgDb();
     $this->dbManager = $this->testDb->getDbManager();
     $userDao = M::mock('Fossology\Lib\Dao\UserDao');
     $uploadDao = M::mock('Fossology\Lib\Dao\UploadDao');
     $this->folderDao = new FolderDao($this->dbManager, $userDao, $uploadDao);
 
-    $this->testDb->createPlainTables(array('folder','foldercontents'));
-    $this->testDb->createSequences(array('folder_folder_pk_seq','foldercontents_foldercontents_pk_seq'));
+    $this->testDb->createPlainTables(array('group_user_member','groups','folder','upload_clearing','foldercontents','upload','uploadtree','license_ref','users'));
+    $this->testDb->createSequences(array('users_user_pk_seq','folder_folder_pk_seq','foldercontents_foldercontents_pk_seq',));
     $this->testDb->createConstraints(array('folder_pkey','foldercontents_pkey'));
     $this->testDb->alterTables(array('folder','foldercontents'));
 
@@ -44,7 +46,38 @@ class FolderDaoTest extends \PHPUnit\Framework\TestCase
     $this->testDb = null;
     $this->dbManager = null;
   }
-
+  /**
+   * @brief Creates a master folder for testing purposes.
+   *
+   * @return int The ID of the newly created folder.
+   */
+  private function createFolder()
+  {
+    $folderId = $this->folderDao->createFolder('Master folder', 'folder for master contents', 1);
+    return $folderId;
+  }
+  /**
+   * @brief Retrieves removable contents of a folder by its ID.
+   *
+   * @param int $folderId The ID of the folder whose contents are to be retrieved.
+   *
+   * @return array List of removable contents.
+   */
+  private function getRemovableContents($folderId)
+  {
+    return $this->folderDao->getRemovableContents($folderId);
+  }
+  /**
+   * @brief Retrieves specific folder content by content ID.
+   *
+   * @param int $folderContentId The ID of the folder content to retrieve.
+   *
+   * @return array|null The folder content or null if not found.
+   */
+  private function getContent($folderContentId)
+  {
+    return $this->folderDao->getContent($folderContentId);
+  }
   public function testGetAllFolderIds()
   {
     $this->testDb->insertData(array('folder'));
@@ -85,6 +118,7 @@ class FolderDaoTest extends \PHPUnit\Framework\TestCase
       array($parentId), __METHOD__);
     assertThat($contentsInfo, is(equalTo(array('foldercontents_mode' => $foldercontentsMode, 'child_id' => $childId))));
   }
+
 
 
   public function testGetFolderPK()
@@ -196,5 +230,104 @@ class FolderDaoTest extends \PHPUnit\Framework\TestCase
     assertThat($goodFolder->getId(), equalTo(FolderDao::TOP_LEVEL));
     $badFolder = $this->folderDao->getFolder(987);
     assertThat($badFolder, is(nullValue()));
+  }
+  /**
+   * @brief Tests the retrieval of the folder tree structure using a Common Table Expression (CTE).
+   */
+  public function testGetFolderTreeCte()
+  {
+    $parent = $this->createFolder();
+    $this->folderDao->createFolder('Child folder', 'child folder example', $parent);
+    $actual = $this->folderDao->getFolderTreeCte($parent);
+    $expected =  'WITH RECURSIVE folder_tree(folder_pk, parent_fk, folder_name, folder_desc, folder_perm, id_path, name_path, depth, cycle_detected) AS (
+  SELECT
+    f.folder_pk, fc.parent_fk, f.folder_name, f.folder_desc, f.folder_perm,
+    ARRAY [f.folder_pk]   AS id_path,
+    ARRAY [f.folder_name] AS name_path,
+    0                     AS depth,
+    FALSE                 AS cycle_detected
+  FROM folder f LEFT JOIN foldercontents fc ON fc.foldercontents_mode=1 AND f.folder_pk=fc.child_id
+  WHERE folder_pk=$1
+  UNION ALL
+  SELECT
+    f.folder_pk, fc.parent_fk, f.folder_name, f.folder_desc, f.folder_perm,
+    id_path || f.folder_pk,
+    name_path || f.folder_name,
+    array_length(id_path, 1),
+    f.folder_pk = ANY (id_path)
+  FROM folder f, foldercontents fc, folder_tree ft
+  WHERE f.folder_pk=fc.child_id AND foldercontents_mode=1 AND fc.parent_fk = ft.folder_pk AND NOT cycle_detected
+)';
+    $this->assertNotNull($actual);
+    $this->assertEquals($expected, $actual);
+  }
+  /**
+   * @brief Tests the retrieval of the folder tree structure using a Common Table Expression (CTE) with null parent ID.
+   */
+  public function testGetFolderTreeCteNullParentId()
+  {
+    $actual = $this->folderDao->getFolderTreeCte();
+    $expected = 'WITH RECURSIVE folder_tree(folder_pk, parent_fk, folder_name, folder_desc, folder_perm, id_path, name_path, depth, cycle_detected) AS (
+  SELECT
+    f.folder_pk, fc.parent_fk, f.folder_name, f.folder_desc, f.folder_perm,
+    ARRAY [f.folder_pk]   AS id_path,
+    ARRAY [f.folder_name] AS name_path,
+    0                     AS depth,
+    FALSE                 AS cycle_detected
+  FROM folder f LEFT JOIN foldercontents fc ON fc.foldercontents_mode=1 AND f.folder_pk=fc.child_id
+  WHERE folder_pk=1
+  UNION ALL
+  SELECT
+    f.folder_pk, fc.parent_fk, f.folder_name, f.folder_desc, f.folder_perm,
+    id_path || f.folder_pk,
+    name_path || f.folder_name,
+    array_length(id_path, 1),
+    f.folder_pk = ANY (id_path)
+  FROM folder f, foldercontents fc, folder_tree ft
+  WHERE f.folder_pk=fc.child_id AND foldercontents_mode=1 AND fc.parent_fk = ft.folder_pk AND NOT cycle_detected
+)';
+    $this->assertNotNull($actual);
+    $this->assertEquals($expected,$actual);
+  }
+  /**
+   * @brief Tests retrieval of folder ID using folder name and parent ID.
+   */
+  public function testGetFolderId()
+  {
+    $parent = $this->createFolder();
+    $expected = $this->folderDao->createFolder("Child folder", "child folder for testing", $parent);
+    $actual = $this->folderDao->getFolderId("Child folder", $parent);
+    $this->assertEquals($expected, $actual);
+  }
+  /**
+   * @brief Tests retrieval of child folders for a given parent folder.
+   */
+  public function testGetChildFolders()
+  {
+    $parentId = $this->createFolder();
+    $this->folderDao->createFolder('App Repo', 'The repository of my app', $parentId);
+    $folders = $this->folderDao->getFolderChildFolders($parentId);
+    var_dump($folders);
+    $this->assertNotNull($folders);
+    $this->assertEquals(1, $folders[2]['foldercontents_mode']);
+  }
+  /**
+   * @brief Tests retrieval of null content when folder content does not exist.
+   */
+  public function testGetFolderContentsIdNullContent()
+  {
+    $result = $this->folderDao->getFolderContentsId(10, 0);
+    $this->assertNull($result);
+  }
+  /**
+   * @brief Tests retrieval of uploads for a specific folder.
+   * @todo Add proper assertions.
+   */
+  public function testGetFolderUploads()
+  {
+    $parent = $this->createFolder();
+    $this->folderDao->createFolder("Child folder", "child folder for testing", $parent);
+    $uploads = $this->folderDao->getFolderUploads($parent);
+    $this->assertNotNull($uploads);
   }
 }
